Fix type error in fbg_fix()
In `fbg_fix()`, `d` has to be of character type, since e.g. ranged values are also saved as a character type within the `baseline_fbg_mgdl` and `updated_fbg_mgdl` columns. if `d` is of numeric type, this would lead to an error. Different types cannot be stored within the same column of a data frame.